requirements.txt
selenium==4.7.2
pandas==1.5.3
requests==2.28.1
google-api-python-client==2.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
